backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pandas
groq
pydantic
